+++ monolith/src/main/java/com/foodie/monolith/model/UserImage.java
@@ -21,6 +21,6 @@ public class UserImage {
 
     @Id
     @Column(name="users_id")
-    private Integer usersId;
+    private Long usersId;
 
 }
