+++ monolith/src/main/java/com/foodie/monolith/repository/UserImageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserImageRepository extends JpaRepository<UserImage, Integer> {
-    Optional<UserImage> findByUsersId(Integer userId);
+    Optional<UserImage> findByUsersId(Long userId);
 
     Optional<UserImage> findByImagesId(Integer imageId);
 }
