+++ monolith/src/main/java/com/foodie/monolith/service/AssignedRoleService.java
@@ -9,10 +9,10 @@ import java.util.Optional;
 public interface AssignedRoleService {
     public List<AssignedRole> getAssignedRoles();
 
-    public Optional<AssignedRole> getAssignedRoleByUserId(Integer userId);
+    public Optional<AssignedRole> getAssignedRoleByUserId(Long userId);
 
     public String assignRoleToUser(AssignedRole newAssignedRole);
 
-    public String deleteAssignedRole(Integer userId);
+    public String deleteAssignedRole(Long userId);
 
 }
